wayland: refactor break out XdgWindow,XdgStatic to xdgwindow.py
